feat: added EXTERNAL_SERVER_URL and INTERNAL_SERVER_URL, renamed folders and files
