+++ .env.sample
@@ -4,3 +4,5 @@ POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
 POSTGRES_DB=postgres
 POSTGRES_PORT=5432
+EXTERNAL_SERVER_URL=http://localhost:3001
+INTERNAL_SERVER_URL=http://server:3001
