--- packages/client/src/constants.ts
@@ -1,6 +1,0 @@
-import './client.d'
-
-export const SERVER_HOST =
-  typeof window === 'undefined'
-    ? __INTERNAL_SERVER_URL__
-    : __EXTERNAL_SERVER_URL__
