+++ packages/client/tsconfig.json
@@ -27,7 +27,7 @@
       "theme": ["./shared/theme/index"],
       "utils": ["./shared/utils/index"],
       "validators": ["./shared/validation/index"],
-      "consts": ["./shared/consts/index"]
+      "constant": ["./shared/constant/index"]
     }
   },
   "include": ["src", "public/sw.ts", "server"],
